backend/requirements.txt
# Зависимости для Pacific.ai Vector Matching Module

# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI API
openai==1.3.7

# Векторные базы данных
pinecone-client==2.2.4
supabase==2.0.2

# База данных
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0

# Обработка переменных окружения
python-dotenv==1.0.0

# Математические вычисления
numpy==1.24.3

# HTTP клиент
httpx==0.25.2

# Логирование и мониторинг
structlog==23.2.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Дополнительные утилиты
python-jose[cryptography]==3.3.0  # Для JWT (опционально)
passlib[bcrypt]==1.7.4  # Для хеширования паролей (опционально)
